requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioop-lts==0.2.2
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.3
fsspec==2026.1.0
gradio==6.3.0
gradio_client==2.0.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
idna==3.11
imbalanced-learn==0.14.1
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
safehttpx==0.1.7
scikit-learn==1.8.0
scipy==1.17.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
starlette==0.50.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
